+++ src/components/ConsortiaControlledVolabulary/index.js
@@ -1,0 +1,1 @@
+export { ConsortiaControlledVolabulary } from './ConsortiaControlledVolabulary';