+++ src/components/index.js
@@ -1,3 +1,4 @@
+export { ConsortiaControlledVolabulary } from './ConsortiaControlledVolabulary';
 export { ConsortiumManagerHeader } from './ConsortiumManagerHeader';
 export { FindConsortiumMember } from './FindConsortiumMember';
 export { SwitchActiveAffiliation } from './SwitchActiveAffiliation';