+++ src/shapes.js
@@ -8,3 +8,19 @@ export const affiliationsShape = PropTypes.arrayOf(PropTypes.shape({
   userId: PropTypes.string,
   username: PropTypes.string,
 }));
+
+export const publicationType = {
+  url: PropTypes.string.isRequired,
+  method: PropTypes.string.isRequired,
+  tenants: PropTypes.arrayOf(PropTypes.string).isRequired,
+  payload: PropTypes.any,
+};
+export const publicationShape = PropTypes.shape(publicationType);
+
+export const translationsShape = PropTypes.shape({
+  cannotDeleteTermHeader: PropTypes.string,
+  cannotDeleteTermMessage: PropTypes.string,
+  deleteEntry: PropTypes.string,
+  termDeleted: PropTypes.string,
+  termWillBeDeleted: PropTypes.string,
+});