+++ src/utils/index.js
@@ -2,3 +2,4 @@ export { checkConsortiumAffiliations } from './checkConsortiumAffiliations';
 export { getCurrentModulePath } from './getCurrentModulePath';
 export { getLegacyTokenHeader } from './getLegacyTokenHeader';
 export { getModuleName } from './getModuleName';
+export { throwErrorResponse } from './throwErrorResponse';