+++ test/jest/fixtures/publishCoordinator.js
@@ -1,0 +1,13 @@
+export const pcPostRequest = {
+  url: '/organizations-storage/organizations',
+  method: 'POST',
+  tenants: [
+    'central',
+    'secondary',
+  ],
+  payload: {
+    name: 'ORG-NAME',
+    status: 'Active',
+    code: 'ORG-CODE',
+  },
+};